+++ src/Client/Driver/DoctrineCacheDriverDecorator.php
@@ -3,9 +3,14 @@
 namespace Gos\Bundle\WebSocketBundle\Client\Driver;
 
 use Doctrine\Common\Cache\Cache;
+use Symfony\Component\Cache\DoctrineProvider;
+
+trigger_deprecation('gos/web-socket-bundle', '3.4', 'The "%s" class is deprecated and will be removed in 4.0, use the "%s" class with a "%s" instance instead.', DoctrineCacheDriverDecorator::class, SymfonyCacheDriverDecorator::class, DoctrineProvider::class);
 
 /**
  * @author Johann Saunier <johann_27@hotmail.fr>
+ *
+ * @deprecated to be removed in 4.0, use the `Gos\Bundle\WebSocketBundle\Client\Driver\SymfonyCacheDriverDecorator` with a `Symfony\Component\Cache\DoctrineProvider` instance instead
  */
 final class DoctrineCacheDriverDecorator implements DriverInterface
 {